refactor in order to run two proxy processes
