+++ src/main.rs
@@ -3,13 +3,10 @@ mod logger;
 mod providers;
 mod proxy;
 
-use hyper::server::conn::http1 as server;
-use hyper::service::service_fn;
-use hyper_util::rt::TokioIo;
-use miette::Result;
-use std::net::SocketAddr;
 use std::sync::Arc;
-use tokio::net::TcpListener;
+
+use providers::Provider;
+use tokio::sync::mpsc::Sender;
 use tracing::{debug, error};
 
 #[derive(Debug)]
@@ -17,8 +14,13 @@ pub enum EventStream {
     PageView(String),
 }
 
+pub struct Platform {
+    pub provider: Arc<Provider>,
+    pub tx: Sender<EventStream>,
+}
+
 #[tokio::main]
-async fn main() -> Result<()> {
+async fn main() {
     let cfg = config::parse();
     logger::init(&cfg.log_severity);
 
@@ -29,28 +31,15 @@ async fn main() -> Result<()> {
         }
     });
 
-    let addr = SocketAddr::from(([127, 0, 0, 1], cfg.http_port));
-    let listener = TcpListener::bind(addr).await.unwrap();
-    tracing::info!(
-        http_port = cfg.http_port,
-        https_port = cfg.https_port,
-        log_severity = cfg.log_severity.as_str(),
-        "Server started"
-    );
+    let platform = Platform {
+        provider: Arc::new(providers::load()),
+        tx,
+    };
 
-    let provider = Arc::new(providers::load());
-
-    loop {
-        let (stream, _) = listener.accept().await.unwrap();
-        let io = TokioIo::new(stream);
-        let provider = Arc::clone(&provider);
-        let proxy = proxy::Proxy::new(provider, tx.clone());
-
-        tokio::task::spawn(async move {
-            server::Builder::new()
-                .serve_connection(io, service_fn(|req| proxy.handle(req)))
-                .await
-                .map_err(|err| error!(%err, "Failed to serve connection"))
-        });
+    tokio::select! {
+        Err(err) = proxy::cleartext::start(cfg.http_port, &platform) => {
+            error!(?err, "HTTP server failed");
+            std::process::exit(1);
+        }
     }
 }
